Revert "Serach seams to be finished, there could be more asserts, concerning routes length and such things, also exec time of binary is not that much shorter than linear"
This reverts commit e7852767eb716cdc4711f31995c96d5437dddc25.

+++ Uebung_1/uebung1/fibonacci_recursive.cpp
@@ -2,7 +2,7 @@
 #include <cstdlib>
 #include <iostream>
 
-unsigned long count = 0;
+int count = 0;
 
 int fibonacci(int number)
 {
